Auth middleware added - GetUser old method commented out for now
